+++ Tests/ConduitTests/Auth/OAuth2TokenGrantManagerTests.swift
@@ -11,7 +11,7 @@ import XCTest
 
 class OAuth2TokenGrantManagerTests: XCTestCase {
 
-    typealias BadResponse = (response: URLResponse?, expectedError: OAuth2Error)
+    typealias BadResponse = (response: HTTPURLResponse?, expectedError: OAuth2Error)
 
     var badResponses: [BadResponse]!