+++ Tests/ConduitTests/Networking/Serialization/XMLNodeTests.swift
@@ -18,6 +18,7 @@ class XMLNodeTests: XCTestCase {
             XMLNode(name: "client", children: [
                 XMLNode(name: "id", value: "client1"),
                 XMLNode(name: "name", value: "Bob"),
+                XMLNode(name: "clientonly", value: "Foo"),
                 XMLNode(name: "customers", children: [
                     XMLNode(name: "customer", children: [
                         XMLNode(name: "id", value: "customer1"),
@@ -32,6 +33,7 @@ class XMLNodeTests: XCTestCase {
             XMLNode(name: "client", children: [
                 XMLNode(name: "id", value: "client2"),
                 XMLNode(name: "name", value: "Job"),
+                XMLNode(name: "clientonly", value: "Bar"),
                 XMLNode(name: "customers", children: [
                     XMLNode(name: "customer", children: [
                         XMLNode(name: "id", value: "customer3"),
@@ -41,7 +43,8 @@ class XMLNodeTests: XCTestCase {
             ]),
             XMLNode(name: "client", children: [
                 XMLNode(name: "id", value: "client3"),
-                XMLNode(name: "name", value: "Joe")
+                XMLNode(name: "name", value: "Joe"),
+                XMLNode(name: "clientonly", value: "Baz")
             ])
         ]),
         XMLNode(name: "id", value: "root1"),
@@ -55,6 +58,7 @@ class XMLNodeTests: XCTestCase {
                 "client": [
                     "id": "client1",
                     "name": "Bob",
+                    "clientonly": "Foo",
                     "customers": [
                         [
                             "customer": [
@@ -75,6 +79,7 @@ class XMLNodeTests: XCTestCase {
                 "client": [
                     "id": "client2",
                     "name": "Job",
+                    "clientonly": "Bar",
                     "customers": [
                         [
                             "customer": [
@@ -88,7 +93,8 @@ class XMLNodeTests: XCTestCase {
             [
                 "client": [
                     "id": "client3",
-                    "name": "Joe"
+                    "name": "Joe",
+                    "clientonly": "Baz"
                 ]
             ]
         ],
@@ -99,6 +105,13 @@ class XMLNodeTests: XCTestCase {
 
     var testSubjects: [XMLNode] {
         return [xml, XMLNode(name: "xml", children: xmlDict)]
+    }
+
+    func testSubscripting() {
+        for subject in testSubjects {
+            XCTAssertEqual(subject["id"]?.value, "root1")
+            XCTAssertNotNil(subject["clients"]?["client"]?["id"])
+        }
     }
 
     func testXMLTreeSearch() {
@@ -112,11 +125,21 @@ class XMLNodeTests: XCTestCase {
         }
     }
 
+    func testXMLFirstLevelSearch() throws {
+        for subject in testSubjects {
+            XCTAssertEqual(subject.nodes(named: "id", traversal: .firstLevel).first?.value, "root1")
+            XCTAssertEqual(try subject.get("name", traversal: .firstLevel), "I'm Root")
+            XCTAssertEqual(try subject.get("rootonly", traversal: .firstLevel), "Root only")
+            XCTAssertThrowsError(try subject.get("clientonly", traversal: .firstLevel) as String)
+        }
+    }
+
     func testXMLDepthTraversal() throws {
         for subject in testSubjects {
             XCTAssertEqual(subject.nodes(named: "id", traversal: .depthFirst).first?.value(), "customer1")
             XCTAssertEqual(try subject.get("name", traversal: .depthFirst), "Customer Awesome")
             XCTAssertEqual(try subject.get("rootonly", traversal: .depthFirst), "Root only")
+            XCTAssertEqual(try subject.get("clientonly", traversal: .depthFirst), "Foo")
         }
     }
 
@@ -125,12 +148,13 @@ class XMLNodeTests: XCTestCase {
             XCTAssertEqual(subject.nodes(named: "id", traversal: .breadthFirst).first?.value(), "root1")
             XCTAssertEqual(try subject.get("name", traversal: .breadthFirst), "I'm Root")
             XCTAssertEqual(try subject.get("rootonly", traversal: .breadthFirst), "Root only")
+            XCTAssertEqual(try subject.get("clientonly", traversal: .breadthFirst), "Foo")
         }
     }
 
     // swiftlint:disable line_length
     func testXMLOutput() {
-        let output = "<xml><clients><client><id>client1</id><name>Bob</name><customers><customer><id>customer1</id><name>Customer Awesome</name></customer><customer><id>customer2</id><name>Another Customer</name></customer></customers></client><client><id>client2</id><name>Job</name><customers><customer><id>customer3</id><name>Yet Another Customer</name></customer></customers></client><client><id>client3</id><name>Joe</name></client></clients><id>root1</id><name>I'm Root</name><rootonly>Root only</rootonly></xml>"
+        let output = "<xml><clients><client><id>client1</id><name>Bob</name><clientonly>Foo</clientonly><customers><customer><id>customer1</id><name>Customer Awesome</name></customer><customer><id>customer2</id><name>Another Customer</name></customer></customers></client><client><id>client2</id><name>Job</name><clientonly>Bar</clientonly><customers><customer><id>customer3</id><name>Yet Another Customer</name></customer></customers></client><client><id>client3</id><name>Joe</name><clientonly>Baz</clientonly></client></clients><id>root1</id><name>I'm Root</name><rootonly>Root only</rootonly></xml>"
         XCTAssertEqual(xml.xmlValue(), output)
     }
     // swiftlint:enable line_length