+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - `XMLNode.children` is no longer optional, defaults to empty array.
 - Auth shared `URLSessionClient` defaults to background operation queue.
 - Rename `XMLNode.`value` property to `textNode`.
+- Update `URLSessionClient` to return `HTTPURLResponse` for easy retrieval of HTTP status codes.
+- Remove `XMLNodeIndex`.
+- Update `XMLNode` subscript method to return `XMLNode?`.
+- Add `.firstLevel` traversal for `XMLNode` to retrieve direct children only.
 
 #### Enhancements
 - Shared `URLSessionClient` with default background operation queue.